fix: use pull request #613 from bdk as crate dependency
Description:
the code was failing when using the standard bdk crate from crate.io.
However, the issue was that bitcoin core version 0.23 does not support
importmulti funnction for descriptor wallets anymore.
This issue was raised in the bdk repo and a potential fix for this was
proposed in pull request #613 by rajarshimaitra.
Now, the code uses the fix from the pull request and it works.